workers/semantics/interface-objects/004.html

<!--
onconnect = function(e) {
  var unexpected = [
    "AbstractView",
    "AbstractWorker",
    "Location",
    "Navigator",
    "DOMImplementation",
    "Audio",
    "HTMLCanvasElement",
    "MouseEvent",
  ];
  var result = [];
  for (var i = 0; i < unexpected.length; ++i) {
    result.push([unexpected[i], unexpected[i] in self]);
  }
  e.ports[0].postMessage(result);
}
/*
-->
<!doctype html>
<title>unavailable interface objects in shared worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
setup(function() {
  window.worker = new SharedWorker('#');
  worker.port.onmessage = function(e) {
    var result = e.data;
    for (var i = 0; i < result.length; ++i) {
      test(function() {
        assert_false(result[i][1]);
      }, "The " + result[i][0] + " interface object should not be exposed");
    }
    done();
  }
}, {explicit_done: true});
</script>
<!--
*/
//-->
